Drop original firmware for MSHW0079
The patched firmware seems to work as intended. As it is a direct
replacement and should not cause any further issues, we can drop the
backup of the original firmware. In case any unforseen issues arise, we
can always get it from the official driver/firmware packages again.

See: https://github.com/jakeday/linux-surface/issues/574